extract graph into separate component and add settings update event
